+++ packages/cli/__tests__/zosfiles/__unit__/list/am/AllMembers.handler.unit.test.ts
@@ -10,7 +10,7 @@
 */
 
 import { List } from "@zowe/zos-files-for-zowe-sdk";
-import { UNIT_TEST_ZOSMF_PROF_OPTS } from "../../../../../../../__tests__/__src__/mocks/ZosmfProfileMock";
+import { UNIT_TEST_ZOSMF_PROF_OPTS } from "../../../../../../../__tests__/__src__/TestConstants";
 import { IHandlerParameters } from "@zowe/imperative";
 
 describe("List AllMembers handler", () => {