+++ packages/cli/__tests__/zosjobs/__unit__/view/all-spool-content/AllSpoolContent.handler.unit.test.ts
@@ -16,7 +16,7 @@ import { GetJobsData } from "../../../__resources__/GetJobsData";
 import { AllSpoolContentDefinition } from "../../../../../src/zosjobs/view/all-spool-content/AllSpoolContent.definition";
 import * as fs from "fs";
 import { ZosmfSession } from "@zowe/zosmf-for-zowe-sdk";
-import { UNIT_TEST_ZOSMF_PROF_OPTS, UNIT_TEST_PROFILES_ZOSMF } from "../../../../../../../__tests__/__src__/mocks/ZosmfProfileMock";
+import { UNIT_TEST_ZOSMF_PROF_OPTS } from "../../../../../../../__tests__/__src__/TestConstants";
 import { mockHandlerParameters } from "@zowe/cli-test-utils";
 import * as AllSpoolContentHandler from "../../../../../src/zosjobs/view/all-spool-content/AllSpoolContent.handler";
 
@@ -27,8 +27,7 @@ const TEST_RESOURCES_DIR = __dirname + "/../../../__resources__";
 const DEFAULT_PARAMETERS: IHandlerParameters = mockHandlerParameters({
     arguments: UNIT_TEST_ZOSMF_PROF_OPTS,
     positionals: ["zos-jobs", "view", "spool-file-by-id"],
-    definition: AllSpoolContentDefinition,
-    profiles: UNIT_TEST_PROFILES_ZOSMF,
+    definition: AllSpoolContentDefinition
 });
 
 describe("zos-jobs view all-spool-content handler", () => {