+++ packages/imperative/src/profiles/src/utils/__tests__/ProfileUtils.unit.test.ts
@@ -16,7 +16,7 @@ import * as path from "path";
 import { IProfileLoaded } from "../../../../index";
 import {
     APPLE_TWO_REQ_DEP_BANANA_ONE_REQ_DEP_GRAPE_ONE_REQ_DEP, BLUEBERRY_PROFILE_TYPE
-} from "../../../../cmd/__tests__/profiles/TestConstants";
+} from "./TestConstants";
 
 const mocks = {
     normalize: path.normalize as unknown as Mock<typeof path.normalize>