+++ __tests__/__packages__/cli-test-utils/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the Zowe CLI test utils package will be documented in this file.
 
+## `8.0.0-next.202408271330`
+
+- BugFix: Removed obsolete V1 `profiles` property from the parameters object returned by `mockHandlerParameters` method.
+
 ## `8.0.0-next.202408131445`
 
 - Update: See `7.28.3` for details